发布文章模块

+++ src/main/java/org/easyblog/bean/Category.java
@@ -7,7 +7,7 @@ public class Category implements Serializable {
 
     private static final long serialVersionUID = 8690952682535328046L;
 
-    private Long categoryId;
+    private int categoryId;
     private Integer categoryUser;
     private String categoryName;
     private String categoryImageUrl;
@@ -38,11 +38,11 @@ public class Category implements Serializable {
         this.createTime = createTime;
     }
 
-    public Long getCategoryId() {
+    public int getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(Long categoryId) {
+    public void setCategoryId(int categoryId) {
         this.categoryId = categoryId;
     }
